+++ src/UX/SapioCompilerModal.js
@@ -1,0 +1,47 @@
+import React from 'react';
+import Button from 'react-bootstrap/Button';
+import Form from 'react-bootstrap/Form';
+import FormControl from 'react-bootstrap/FormControl';
+import Modal from 'react-bootstrap/Modal';
+
+export class SapioCompilerModal extends React.Component {
+    constructor(props) {
+        super(props);
+        this.form = {};
+    }
+    handleSubmit(event) {
+        event.preventDefault();
+        event.stopPropagation();
+        const form = event.currentTarget;
+        this.props.compiler.location = form.elements.ws.value;
+        // triggers reconnect
+        this.props.compiler.socket.close();
+        this.props.hide();
+    }
+    render() {
+        return (
+            <Modal show={this.props.show} onHide={this.props.hide}>
+                <Modal.Header closeButton>
+                    <Modal.Title> Set Contract Generator URL </Modal.Title>
+                </Modal.Header>
+                <Form onSubmit={(e) => this.handleSubmit(e)}>
+                    <FormControl
+                        name="ws"
+                        type="text"
+                        placeholder="url"
+                        defaultValue={this.props.compiler.location} />
+                    <Button type="submit">Set</Button>
+                </Form>
+                <Modal.Footer>
+                    <Button
+                        variant="secondary"
+                        onClick={() => this.props.hide()}
+                    >
+                        {' '}
+                        Close{' '}
+                    </Button>
+                </Modal.Footer>
+            </Modal>
+        );
+    }
+}
